+++ lib/preload_google_ads.dart
@@ -17,4 +17,5 @@ export 'src/app_open_ad/on_demand_app_open_ad.dart';
 export 'src/banner_ad/show_collapsible_banner_ad.dart';
 export 'src/inter_ad/on_demand_interstitial_ad.dart';
 export 'src/native_ad/native_ad_style.dart';
+export 'src/rewarded_ad/on_demand_rewarded_ad.dart';
 export 'src/rewarded_inter_ad/on_demand_rewarded_inter_ad.dart';